Address svojsu review: dedupe staking-warning presentation via StakingWarningPresentable
New protocol + default-impl extension holds the modal-sheet plumbing that
was previously duplicated in BaseDAppBrowserWireframe.presentStakingWarning
and BrowserNavigationTaskFactory.presentStakingWarning. This commit adopts
it in BaseDAppBrowserWireframe; the pre-browser path adopts it in the
follow-up commit.

Follows the existing AlertPresentable / ErrorPresentable / DAppAlertPresentable
convention in the codebase.
